Subjective_Outcomes_OpenAI/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.0
ansi2html==1.8.0
anyio==3.6.2
asgiref==3.6.0
aspose-words==23.1.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.9.3
blinker==1.5
blis==0.7.9
bs4==0.0.1
build==0.9.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
comtypes==1.1.14
confection==0.0.3
cryptography==3.4.6
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==4.4
debugpy==1.6.7
decorator==4.4.2
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
dnspython==1.16.0
docx2pdf==0.1.8
docx2txt==0.8
email-validator==1.3.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
en-core-web-sm==3.1.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.84.0
fastjsonschema==2.16.3
feedfinder2==0.0.4
feedparser==6.0.10
ffmpeg-python==0.2.0
filelock==3.9.0
flashtext==2.7
Flask==2.0.2
fonttools==4.39.2
fpdf==1.7.2
freetype-py==2.3.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
gnews==0.2.7
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.11.1
idna==3.3
imageio==2.28.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
itsdangerous==2.0.1
jedi==0.18.2
jieba3k==0.35.1
Jinja2==3.0.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kaleido==0.2.1
kiwisolver==1.4.4
langcodes==3.3.0
lightning-utilities==0.6.0.post0
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
multipart==0.2.4
multiprocess==0.70.14
multitasking==0.0.11
murmurhash==1.0.9
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
networkx==2.8.8
newspaper3k==0.2.8
nltk==3.7
numba==0.56.4
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.5
openai-whisper==20230117
openpyxl==3.0.10
orjson==3.8.8
packaging==21.3
pandas==1.4.2
pandas-stubs==1.2.0.35
parso==0.8.3
pathos==0.3.0
pathy==0.10.1
patsy==0.5.3
pdfminer==20191125
pdfservices-sdk==2.1.2
pep517==0.13.0
pickleshare==0.7.5
Pillow==9.3.0
pke==2.0.0
platformdirs==2.6.2
plotly==5.6.0
polling==0.3.2
polling2==0.5.0
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.8
proglog==0.1.10
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycairo==1.23.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
pydeck==0.8.0
PyDrive==1.3.1
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.14.0
PyJWT==2.4.0
pymongo==3.12.3
Pympler==1.0.1
pyodbc==4.0.34
pyparsing==3.0.9
PyPDF2==2.12.1
pyperclip==1.8.2
pypng==0.20220715.0
PyQRCode==1.2.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytorch-lightning==1.9.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==300
PyYAML==6.0
pyzmq==25.0.2
ray==2.2.0
regex==2022.3.15
reportlab==4.0.0
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rich==13.3.2
rlPyCairo==0.2.0
rsa==4.9
Sastrawi==1.0.1
schemdraw==0.16
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.0
semver==2.13.0
sense2vec==2.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0.post1
sktime==0.17.1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
srsly==2.4.5
stack-data==0.6.2
starlette==0.19.1
statsmodels==0.13.5
streamlit==1.11.1
streamlit-option-menu==0.3.2
sympy==1.11.1
tenacity==8.2.2
tensorboardX==2.5.1
thinc==8.1.7
threadpoolctl==3.1.0
tika==2.6.0
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchmetrics==0.11.1
torchvision==0.15.1
tornado==6.2
tqdm==4.62.3
trace-updater==0.0.9.1
traitlets==5.9.0
transformers==4.26.1
translators==5.0.0
twython==3.8.0
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.3
ujson==5.7.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.17.6
vaderSentiment==3.3.2
validators==0.20.0
virtualenv==20.17.1
wasabi==1.1.1
watchdog==3.0.0
watchgod==0.8.2
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.0.2
widgetsnbextension==4.0.7
win32-setctime==1.1.0
wrapt==1.15.0
xxhash==3.2.0
yarl==1.8.2
yellowbrick==1.5
yfinance==0.1.70
zipp==3.15.0
zope.interface==6.0
